git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/simulations/mw@12253 6e01202a-0783-4428-890a-84243c50cc2b

+++ src/org/concord/modeler/g2d/Curve.java
@@ -163,7 +163,7 @@ public class Curve implements Serializable {
 					tmp = 0;
 					for (int p = i - k; p < i; p++)
 						tmp += yQ.getData(p);
-					tmp /= (float) k;
+					tmp /= k;
 				}
 				break;
 			case EXPONENTIAL_RUNNING_AVERAGE:
